It works now, not cleaned up
Fixed an indexing issue that was occuring when it was "categories[i]", instead of "data.get(i).category". I was essentially only checking a category against itself, instead of the correct category. So what was happening was every answer node was being assigned a correct answer of 1, instead of 0 and 1